backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.3
python-jose[cryptography]>=3.3.0
pydantic[email-validator]>=2.9.0
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
